create checkout component
